Restructured the data structure coming _out_ of the Store from getTabs to align with the intended use by the UI. For example, axes along which we need stable ordering are now arrays, not dictionaries. This enables easy display of table information, and I have adjusted the ListController to match.  It now shows the tabs for each client in a separate section of the table, labeled by client.
This refactoring ought to be done for each of the other data types.
In addition, there should really be a different UITableView object for each of the tables we have, instead of one that keeps throwing away its contents, and is full of "if we're showing tabs, do this, else if we're showing history, do this" , which is the opposite of object-oriented.